Update gradle, refactor BlockRendererRegistry, implement Forgix plugin
The core changes include the update to gradle version from 8.4 to 8.6, introduction of variation handling in BlockRendererRegistry, and implementation of the Forgix plugin in build.gradle for merging mods. From the broader perspective, these changes are intended to improve compatibility, performance, and flexibility of the mod. A new gradle-wrapper.properties setting for validateDistributionUrl has been added for additional security.