Update door animation and remove Russian translation
Modified the door animation in "DoorRenderer.java" to accommodate both left and right hinged doors, enhancing the user experience. The Russian translation file "ru_ru.json" was removed because it is no longer necessary. Also, a minor amendment was made to "fabric.mod.json" - the "badges" line was deleted. The mesh transformation process in "Mesh.java" was revised, notably comments were added for future improvements, and the type of globalMatrix was changed from Matrix4f to Matrix4fc to increase precision.