Refactor build files and update parts of the codebase
This commit includes restructuring of the build files, specifically the build.gradle files, to better organize dependencies. Additionally, several sections of the codebase have been updated for clarity and performance. This includes enhancements to the Animation class and changes to the MeshBlockRenderer and PLRenderType classes. Some unnecessary imports and configurations were also removed.